Modal working fine before lib integraton
