--- src/App.css
@@ -1,34 +1,0 @@
-.App {
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
-
-body {
-  margin: 0;
-  padding: 0;
-}
-
-.app-container {
-  display: grid;
-  grid-template-columns: max-content 1fr;
-  grid-template-rows: 1fr;
-  gap: 0px 0px;
-  grid-template-areas:
-    "sideBar  Main";
-  height: 100vh;
-}
-
-.font-face-gm {
-  font-family: "times";
-}
-
-.main-container {
-  height: 90%;
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-  padding-top: 4rem;
-
-}
