+++ src/Components/SideNav/sideNav.scss
@@ -19,7 +19,7 @@
     font-size: 1rem;
     height: 100%;
 }
-    /* Styles pour les petits écrans */
+   // Styles pour les petits écrans 
     @media (max-width: $small-screen) {
         .main-nav {@include side-nav-mobile;
             height: 5%;
