+++ src/Components/Table/AllEmployeesTable.scss
@@ -1,4 +1,4 @@
-@import '../../styles/main';
+@import "../../styles/main"; 
 
 .employees-table-container{
